+++ src/Acl.Fs.Core/Abstractions/Factory/IXChaCha20Poly1305Factory.cs
@@ -1,0 +1,8 @@
+﻿using NSec.Cryptography;
+
+namespace Acl.Fs.Core.Abstractions.Factory;
+
+internal interface IXChaCha20Poly1305Factory
+{
+    XChaCha20Poly1305 Create(byte[] key);
+}