+++ src/Acl.Fs.Core/Extensions/Encryption/ServiceExtensions.cs
@@ -1,9 +1,11 @@
 ﻿using System.Security.Cryptography;
+using Acl.Fs.Core.Abstractions.Factory;
 using Acl.Fs.Core.Abstractions.Service.Encryption.AesGcm;
 using Acl.Fs.Core.Abstractions.Service.Encryption.Shared.Audit;
 using Acl.Fs.Core.Abstractions.Service.Encryption.Shared.Metadata;
 using Acl.Fs.Core.Abstractions.Service.Encryption.Shared.Processor;
 using Acl.Fs.Core.Abstractions.Service.Encryption.Shared.Validation;
+using Acl.Fs.Core.Factory;
 using Acl.Fs.Core.Service.Encryption.AesGcm;
 using Acl.Fs.Core.Service.Encryption.Shared.Audit;
 using Acl.Fs.Core.Service.Encryption.Shared.Metadata;
@@ -12,6 +14,8 @@ using Acl.Fs.Core.Service.Encryption.Shared.Provider;
 using Acl.Fs.Core.Service.Encryption.Shared.Validation;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using NSec.Cryptography;
+using ChaCha20Poly1305 = System.Security.Cryptography.ChaCha20Poly1305;
 
 namespace Acl.Fs.Core.Extensions.Encryption;
 
@@ -49,6 +53,22 @@ public static class ServiceExtensions
         services
             .TryAddScoped<Abstractions.Service.Encryption.ChaCha20Poly1305.IEncryptionService,
                 Service.Encryption.ChaCha20Poly1305.EncryptionService>();
+
+        return services;
+    }
+
+    public static IServiceCollection AddXChaCha20Poly1305EncryptionServices(this IServiceCollection services)
+    {
+        services.TryAddScoped<IXChaCha20Poly1305Factory, XChaCha20Poly1305Factory>();
+        services.TryAddScoped<ICryptoProvider<Key>, XChaCha20Poly1305CryptoProvider>();
+        services.TryAddScoped<IBlockProcessor<Key>, BlockProcessor<Key>>();
+
+        services
+            .TryAddScoped<Abstractions.Service.Encryption.XChaCha20Poly1305.IEncryptorBase,
+                Service.Encryption.XChaCha20Poly1305.EncryptorBase>();
+        services
+            .TryAddScoped<Abstractions.Service.Encryption.XChaCha20Poly1305.IEncryptionService,
+                Service.Encryption.XChaCha20Poly1305.EncryptionService>();
 
         return services;
     }