+++ src/Acl.Fs.Core/Service/Encryption/Shared/Buffer/BufferManager.cs
@@ -5,13 +5,13 @@ using static Acl.Fs.Constant.Cryptography.CryptoConstants;
 
 namespace Acl.Fs.Core.Service.Encryption.Shared.Buffer;
 
-internal sealed class BufferManager(int metadataBufferSize) : IBufferManager
+internal sealed class BufferManager(int metadataBufferSize, int nonceSize) : IBufferManager
 {
     public byte[] Buffer { get; } = CryptoPool.Rent(BufferSize);
     public byte[] Ciphertext { get; } = CryptoPool.Rent(BufferSize);
     public byte[] MetadataBuffer { get; } = CryptoPool.Rent(metadataBufferSize);
     public byte[] Tag { get; } = CryptoPool.Rent(TagSize);
-    public byte[] ChunkNonce { get; } = CryptoPool.Rent(NonceSize);
+    public byte[] ChunkNonce { get; } = CryptoPool.Rent(nonceSize);
     public byte[] Salt { get; } = CryptoPool.Rent(SaltSize);
 
     public void Dispose()