+++ src/Acl.Fs.Core/Versioning/ValidationStrategies/V0ValidationStrategy.cs
@@ -9,7 +9,9 @@ internal sealed class V0ValidationStrategy : IVersionValidationStrategy
 {
     public void Validate(byte minorVersion)
     {
-        // For beta version 0.x, only minor version 1 is currently supported
+        // For beta version 0.x, minor versions 1 and 2 are currently supported
+        // v0.1.x: Initial release with AES-GCM and ChaCha20Poly1305
+        // v0.2.x: release with XChaCha20Poly1305 support
         if (minorVersion > VersionConstants.CurrentMinorVersion)
             throw new VersionValidationException(
                 string.Format(ErrorMessages.FutureMinorVersionNotSupported,